requirements.txt
scipy
# wandb
colorlog
tqdm
h5py
ipdb
opencv-python
moviepy
mpi4py
gym
mujoco-py
absl-py
gdown
decorator==4.0.2
numpy-quaternion
transforms3d
numba
